Renamed to error-strings.h
