+++ storage/ndb/rest-server/data-access-rondb/src/db-operations/pk/common.cpp
@@ -26,7 +26,7 @@
 #include <string>
 #include <algorithm>
 #include <utility>
-#include "src/error-strs.h"
+#include "src/error-strings.h"
 #include "src/status.hpp"
 #include "src/mystring.hpp"
 #include "src/rdrs-const.h"
